% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.linreg}
\alias{as.data.frame.linreg}
\title{A function that creates a data frame from an object of class linreg.}
\usage{
\method{as.data.frame}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other arguments}
}
\value{
data.frame contains fitted values and residuals.
}
\description{
A function that creates a data frame from an object of class linreg.
}
