% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{package}
\name{boxlinre}
\alias{boxlinre}
\alias{boxlinre-package}
\title{boxlinre: A linear regression package in R}
\description{
A linear regression package in R
}
\references{
\url{https://en.wikipedia.org/wiki/Linear_regression}
}
\author{
Maintainer: Boxi Zhang bossyzhang@gmail.com
}
