% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{A method that returns the coefficients as a named vector.}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{other arguments}
}
\value{
named vector of coefficients.
}
