% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef}
\alias{coef}
\title{A method that returns the coefficients as a named vector}
\usage{
coef(x, ...)
}
\arguments{
\item{...}{other arguments}

\item{object}{linreg object}
}
\value{
named vector of coefficients.
}
\description{
A method that returns the coefficients as a named vector
}
