% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Multiple Linear Regression.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula implies the model (e.g. y ~ x).}

\item{data}{data frame, dataset attached to the algorithm.}
}
\value{
a list.
}
\description{
Multiple Linear Regression.
}
\examples{
linreg(eruptions~waiting, faithful)
}
