% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{The plot-method that creates two plots}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other arguments}
}
\value{
ggplot2 figures
}
\description{
The plot-method that creates two plots
}
