% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pred}
\alias{pred}
\title{A method that returns the predicted values y.}
\usage{
pred(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other arguments}
}
\value{
vector of predictions.
}
\description{
A function that return the predicted values y.
}
