% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Method for printing the coefficients and coefficient names}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other arguments}
}
\description{
Method for printing the coefficients and coefficient names
}
