% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.linreg_summary}
\alias{print.linreg_summary}
\title{Method for printing a summary of the results from the linreg function}
\usage{
\method{print}{linreg_summary}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other arguments}
}
\description{
Method for printing a summary of the results from the linreg function
}
