% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{resid}
\alias{resid}
\title{A method that returns the vector of residuals}
\usage{
resid(x, ...)
}
\arguments{
\item{x}{linreg object.}

\item{...}{other arguments.}
}
\value{
vector of residuals.
}
\description{
A function that check the residulas from an object of class linreg
}
