% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{A method that creates a summary of the linreg-function.}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{other arguments}
}
\value{
linreg_summary object, can be printed.
}
