% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary}
\alias{summary}
\title{A method that creates a summary of the linreg-function}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other arguments}
}
\value{
linreg_summary object, can be printed.
}
\description{
A method that creates a summary of the linreg-function
}
